Simple fix for github issue #3
This is a simple and not very clean fix for
gridcf/UberFTP/issues/3
Either (one of) the l_write() or the l_close() can fail if the remote file
already exists, since the data goes over a different channel than the control
messages. For large files, it will typically be one of the later l_write()
statements, for small files probably the l_close(). Hence we effectively only
have the errmsg to find out what happened, and therefore do a simple str(n)cmp
with this errmsg. For this we also need to add a function to get it from the
opaque errcode_t.